moved strats to constants
